Move shared code to base class
